pyproject.toml
[project]
name = "datascienceproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib==1.5.1",
    "numpy==2.3.1",
    "python-box==7.3.2",
    "pyyaml==6.0.2",
]
